+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.2.1] -
+### Fixed
+- Approximate model counting subprocess newlines and encoding
+- `set_type` docstring
+- `props.influence` docstring
+- Ternary transform for circuits with xor/xnor gates
+- `sat.construct_solver` works with newer versions of `python-sat`
+
+### Added
+- Generic flop blackbox
+- `insert_registers` transform
+- `limit_fanout` transform
+- `levelize` function
+
+### Changed
+- `sat.construct_solver` accepts a `pysat.Solver` class instead of a string for greater flexibility. If no solver is specified, `Cadical` (`Cadical153` for newer versions of `python-sat`) will be used.
+- Added `stretch` argument to Yosys `show` command for `cg.visualize`
+
+
 ## [0.2.0] - 2022-04-22
 ### Fixed
 - Visualization with BlackBoxes.